OpenCL initialization routine
